+++ lib/src/process_unique/process_unique.dart
@@ -2,7 +2,7 @@ import './process_unique_fallback.dart'
     if (dart.library.html) './process_unique_web.dart'
     if (dart.library.io) './process_unique_io.dart';
 
-/// Process unique astract class
+/// Process unique abstract class
 abstract class ProcessUnique {
   /// Get process unique random value
   int getValue();